+++ crates/futures/Cargo.toml
@@ -22,4 +22,4 @@ lazy_static = { version = "1.3.0", optional = true }
 wasm-bindgen-test = { path = '../test', version = '0.2.43' }
 
 [features]
-nightly = ["futures-util-preview", "futures-channel-preview", "lazy_static"]
+futures_0_3 = ["futures-util-preview", "futures-channel-preview", "lazy_static"]
